patch: 0.1.1 后端代码提交

+++ src/backend/bisheng/api/v1/flows.py
@@ -7,9 +7,12 @@ from bisheng.api.v1.schemas import FlowListCreate, FlowListRead
 from bisheng.database.base import get_session
 from bisheng.database.models.flow import (Flow, FlowCreate, FlowRead,
                                           FlowReadWithStyle, FlowUpdate)
+from bisheng.database.models.template import Template
+from bisheng.database.models.user import User
 from bisheng.settings import settings
 from fastapi import APIRouter, Depends, File, HTTPException, Query, UploadFile
 from fastapi.encoders import jsonable_encoder
+from fastapi_jwt_auth import AuthJWT
 from sqlmodel import Session, select
 
 # build router
@@ -17,12 +20,16 @@ router = APIRouter(prefix='/flows', tags=['Flows'])
 
 
 @router.post('/', response_model=FlowRead, status_code=201)
-def create_flow(*, session: Session = Depends(get_session), flow: FlowCreate):
+def create_flow(*, session: Session = Depends(get_session), flow: FlowCreate, Authorize: AuthJWT = Depends()):
     """Create a new flow."""
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
+
     if flow.flow_id:
         # copy from template
         temp_flow = session.get(Flow, flow.flow_id)
         flow.data = temp_flow.data
+    flow.user_id = payload.get('user_id')
     db_flow = Flow.from_orm(flow)
     session.add(db_flow)
     session.commit()
@@ -31,17 +38,21 @@ def create_flow(*, session: Session = Depends(get_session), flow: FlowCreate):
 
 
 @router.get('/', response_model=list[FlowReadWithStyle], status_code=200)
-def read_flows(
-    *,
-    session: Session = Depends(get_session),
-    name: str = Query(default=None, description='根据name查找数据库'),
-    page_size: int = Query(default=None, description='根据pagesize查找数据库'),
-    page_num: int = Query(default=None, description='根据pagenum查找数据库'),
-    status: int = None
-):
+def read_flows(*,
+               session: Session = Depends(get_session),
+               name: str = Query(default=None, description='根据name查找数据库'),
+               page_size: int = Query(default=None, description='根据pagesize查找数据库'),
+               page_num: int = Query(default=None, description='根据pagenum查找数据库'),
+               status: int = None,
+               Authorize: AuthJWT = Depends()):
     """Read all flows."""
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
+
     try:
         sql = select(Flow)
+        if 'admin' != payload.get('role'):
+            sql = sql.where(Flow.user_id == payload.get('user_id'))
         if name:
             sql = sql.where(Flow.name.like(f'%{name}%'))
         if status:
@@ -49,10 +60,19 @@ def read_flows(
 
         sql = sql.order_by(Flow.update_time.desc())
         if page_num and page_size:
-            sql = sql.offset((page_num-1) * page_size).limit(page_size)
+            sql = sql.offset((page_num - 1) * page_size).limit(page_size)
 
         flows = session.exec(sql).all()
-        return [jsonable_encoder(flow) for flow in flows]
+
+        res = [jsonable_encoder(flow) for flow in flows]
+        if flows:
+            db_user_ids = {flow.user_id for flow in flows}
+            db_user = session.exec(select(User).where(User.user_id.in_(db_user_ids))).all()
+            userMap = {user.user_id: user.user_name for user in db_user}
+            for r in res:
+                r['user_name'] = userMap[r['user_id']]
+
+        return res
 
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e)) from e
@@ -68,26 +88,24 @@ def read_flow(*, session: Session = Depends(get_session), flow_id: UUID):
 
 
 @router.patch('/{flow_id}', response_model=FlowRead, status_code=200)
-def update_flow(
-    *,
-    session: Session = Depends(get_session),
-    flow_id: UUID,
-    flow: FlowUpdate
-):
+def update_flow(*, session: Session = Depends(get_session), flow_id: UUID, flow: FlowUpdate, Authorize: AuthJWT = Depends()):
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
     """Update a flow."""
     db_flow = session.get(Flow, flow_id)
     if not db_flow:
         raise HTTPException(status_code=404, detail='Flow not found')
+
+    if 'admin' != payload.get('role') and db_flow.user_id != payload.get('user_id'):
+        raise HTTPException(status_code=500, detail='没有权限编辑此技能')
+
     flow_data = flow.dict(exclude_unset=True)
 
-    if 'status' in flow_data and flow_data['status'
-                                           ] == 2 and db_flow.status == 1:
+    if 'status' in flow_data and flow_data['status'] == 2 and db_flow.status == 1:
         # 上线校验
         try:
             art = {}
-            build_flow_no_yield(
-                graph_data=db_flow.data, artifacts=art, process_file=False
-            )
+            build_flow_no_yield(graph_data=db_flow.data, artifacts=art, process_file=False)
         except Exception as exc:
             raise HTTPException(status_code=500, detail='Flow 编译不通过') from exc
 
@@ -102,11 +120,21 @@ def update_flow(
 
 
 @router.delete('/{flow_id}', status_code=200)
-def delete_flow(*, session: Session = Depends(get_session), flow_id: UUID):
+def delete_flow(*, session: Session = Depends(get_session), flow_id: UUID, Authorize: AuthJWT = Depends()):
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
     """Delete a flow."""
     flow = session.get(Flow, flow_id)
     if not flow:
         raise HTTPException(status_code=404, detail='Flow not found')
+    if 'admin' != payload.get('role') and flow.user_id != payload.get('user_id'):
+        raise HTTPException(status_code=500, detail='没有权限删除此技能')
+
+    # 判断是否属于模板
+    db_template = session.exec(select(Template).where(Template.flow_id == flow_id)).first()
+    if db_template:
+        session.delete(db_template)
+
     session.delete(flow)
     session.commit()
     return {'message': 'Flow deleted successfully'}
@@ -114,13 +142,14 @@ def delete_flow(*, session: Session = Depends(get_session), flow_id: UUID):
 
 # Define a new model to handle multiple flows
 @router.post('/batch/', response_model=List[FlowRead], status_code=201)
-def create_flows(
-    *, session: Session = Depends(get_session), flow_list: FlowListCreate
-):
+def create_flows(*, session: Session = Depends(get_session), flow_list: FlowListCreate, Authorize: AuthJWT = Depends()):
+    Authorize.jwt_required()
+    payload = json.loads(Authorize.get_jwt_subject())
     """Create multiple new flows."""
     db_flows = []
     for flow in flow_list.flows:
         db_flow = Flow.from_orm(flow)
+        db_flow.user_id = payload.get('user_id')
         session.add(db_flow)
         db_flows.append(db_flow)
     session.commit()
@@ -130,9 +159,7 @@ def create_flows(
 
 
 @router.post('/upload/', response_model=List[FlowRead], status_code=201)
-async def upload_file(
-    *, session: Session = Depends(get_session), file: UploadFile = File(...)
-):
+async def upload_file(*, session: Session = Depends(get_session), file: UploadFile = File(...), Authorize: AuthJWT = Depends()):
     """Upload flows from a file."""
     contents = await file.read()
     data = json.loads(contents)
@@ -140,7 +167,7 @@ async def upload_file(
         flow_list = FlowListCreate(**data)
     else:
         flow_list = FlowListCreate(flows=[FlowCreate(**flow) for flow in data])
-    return create_flows(session=session, flow_list=flow_list)
+    return create_flows(session=session, flow_list=flow_list, Authorize=Authorize)
 
 
 @router.get('/download/', response_model=FlowListRead, status_code=200)
